Refactor env discovery to the web3 package
historically this is where the deployments live, and this will break less code in harmony as well.
